+++ src/ae_kqueue.c
@@ -29,36 +29,16 @@
  */
 
 
+#include "file_opt_unix.h"
+
 #include <sys/types.h>
 #include <sys/event.h>
 #include <sys/time.h>
-#include <fcntl.h>  /* O_CLOEXEC */
-#include <errno.h>
 
 typedef struct aeApiState {
     int kqfd;
     struct kevent *events;
 } aeApiState;
-
-int cloexecFcntl(int fd) {
-    int r;
-    int flags;
-
-    do
-        r = fcntl(fd, F_GETFD);
-    while (r == -1 && errno == EINTR);
-
-    if (r == -1 || (r & FD_CLOEXEC))
-        return r;
-
-    flags = r | FD_CLOEXEC;
-
-    do
-        r = fcntl(fd, F_SETFD, flags);
-    while (r == -1 && errno == EINTR);
-
-    return r;
-}
 
 static int aeApiCreate(aeEventLoop *eventLoop) {
     aeApiState *state = zmalloc(sizeof(aeApiState));
